Create README.txt
My journey into the computer world.

By the beginning of the year 2023, I was a computer dimwit per say. My computer literacy skills were most rudimentary and primitive. Having grown up in an African countryside where technological exposure till date is low, a desire to get into tech is tantamount to a wild goose chase. It calls for unmatched resilience and commitment to get a breakthrough.

Early this year, I met with an old friend ( We first met when I joined the high school where he was a student.), who is doing well in the field of tech. He had these mind-blowing idea of a project he wanted to roll-out. He explained in a nutshell what the process would look like. I listened keenly but at first thought I could not be of help, since I had no requisite skills.

My friend told me that it was not a prerequisite to have computer skills. The major requirement was to be committed to the course and be teachable. I took up the challenge to learn new skills and be productive in society. Three months now into this thrilling journey and I cannot find enough words to thank all my mentors. I am also open to collaborate with everyone with a similar mindset to continue scaling the heights.

The bigger picture.
I am striving to measure up to the purpose. It is in the public domain that knowledge is power. I am one very happy person for finally having found a purpose in life. I look forward to a point in time when I will be in a position to mentor other soul(s) to live a purposeful live and be that light at the end of their tunnels.